Fix SvgCircle

+++ src/MudBlazor/Components/Chart/Svg/SvgCircle.cs
@@ -3,10 +3,10 @@ namespace MudBlazor.Charts.SVG.Models
     public class SvgCircle
     {
         public int Index { get; set; }
-        public decimal CX { get; set; }
-        public decimal CY { get; set; }
-        public decimal Radius { get; set; }
+        public double CX { get; set; }
+        public double CY { get; set; }
+        public double Radius { get; set; }
         public string StrokeDashArray { get; set; }
-        public decimal StrokeDashOffset { get; set; }
+        public double StrokeDashOffset { get; set; }
     }
 }
